feat(syntax): integrate tree-sitter incremental parsing, code highlighter, outline extractor, and Stage 29 CI
- Implement TTZipSyntaxParser GLR incremental parser with subtree reuse
- Implement SyntaxHighlighter with Tree-sitter S-expression query pattern matching
- Implement SymbolOutlineExtractor for zero-allocation cross-language AST outlines
- Implement LanguageRegistry with automatic extension and shebang heuristics
- Implement Swift 6 TTZipSyntaxHighlightService facade with actor caching and UniFFI 0.28 bindings
- Deploy 6-layer syntax defense-in-depth against deep AST recursion and GLR loop bombs
- Upgrade run_local_ci_gate.sh to 29 stages (all 29 stages green in ~220s)

+++ rust/ttzip-engine/tests/zlib_ng_fuzz_tests.rs
@@ -273,7 +273,7 @@ fn test_target_06_1000_task_concurrent_dual_engine_arbitration() {
             *b = prng.next_u8();
         }
 
-        let level = (task_id % 13) as i32;
+        let level = task_id % 13;
 
         // 1. TTZip Compress -> flate2 Decompress (Zlib)
         let bound = zlib_compress_bound(raw.len(), level);
